requirements_lin.txt
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
click==8.1.8
colorama==0.4.6
diff-match-patch==20241021
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-import-export==4.3.4
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.0.8
rq==1.16.2
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
times==0.7
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
